refactor(pool): split acquire route versus buckets
